@@ -1,46 +1,57 @@
 {
-  "name": "client",
-  "version": "1.0.0",
-  "scripts": {
-    "build": "webpack --mode production",
-    "build:dev": "webpack --mode development",
-    "build:start": "cd dist && PORT=3000 npx serve",
-    "start": "webpack serve --open --mode development",
-    "start:live": "webpack serve --open --mode development --live-reload --hot"
-  },
-  "license": "MIT",
-  "author": {
-    "name": "Jack Herrington",
-    "email": "jherr@pobox.com"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.15.8",
-    "@babel/plugin-transform-runtime": "^7.15.8",
-    "@babel/preset-env": "^7.15.8",
-    "@babel/preset-react": "^7.14.5",
-    "@babel/preset-typescript": "^7.10.4",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.2",
-    "autoprefixer": "^10.1.0",
-    "babel-loader": "^8.2.2",
-    "css-loader": "^6.3.0",
-    "html-webpack-plugin": "^5.3.2",
-    "postcss": "^8.2.1",
-    "postcss-loader": "^4.1.0",
-    "style-loader": "^3.3.0",
-    "tailwindcss": "^2.0.2",
-    "typescript": "^4.5.2",
-    "webpack": "^5.57.1",
-    "webpack-cli": "^4.9.0",
-    "webpack-dev-server": "^4.3.1"
-  },
-  "dependencies": {
-    "@tanstack/react-query": "^4.23.0",
-    "@trpc/client": "^10.9.0",
-    "@trpc/react-query": "^10.9.0",
-    "@trpc/server": "^10.9.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
-  },
-  "proxy": "http://localhost:8080/"
+	"name": "client",
+	"version": "1.0.0",
+	"scripts": {
+		"build": "webpack --mode production",
+		"build:dev": "webpack --mode development",
+		"build:start": "cd dist && PORT=3000 npx serve",
+		"start": "webpack serve --open --mode development",
+		"start:live": "webpack serve --open --mode development --live-reload --hot"
+	},
+	"license": "MIT",
+	"author": {
+		"name": "Chase Buchholz",
+		"email": "cdbuchholz10@gmail.com"
+	},
+	"devDependencies": {
+		"@babel/core": "^7.15.8",
+		"@babel/plugin-transform-runtime": "^7.15.8",
+		"@babel/preset-env": "^7.15.8",
+		"@babel/preset-react": "^7.14.5",
+		"@babel/preset-typescript": "^7.10.4",
+		"@types/react": "^18.0.10",
+		"@types/react-dom": "^18.0.10",
+		"@types/tailwindcss": "^3.1.0",
+		"@typescript-eslint/eslint-plugin": "^5.50.0",
+		"@typescript-eslint/parser": "^5.50.0",
+		"autoprefixer": "^10.1.0",
+		"babel-loader": "^8.2.2",
+		"css-loader": "^6.3.0",
+		"eslint": "^8.33.0",
+		"eslint-plugin-react": "^7.32.2",
+		"html-webpack-plugin": "^5.3.2",
+		"postcss": "^8.2.1",
+		"postcss-loader": "^4.1.0",
+		"style-loader": "^3.3.0",
+		"tailwindcss": "^2.0.2",
+		"typescript": "^4.5.2",
+		"url-loader": "^4.1.1",
+		"webpack": "^5.57.1",
+		"webpack-cli": "^4.9.0",
+		"webpack-dev-server": "^4.3.1"
+	},
+	"dependencies": {
+		"@emotion/react": "^11.10.5",
+		"@mantine/core": "^5.10.3",
+		"@mantine/hooks": "^5.10.3",
+		"@tabler/icons-react": "^2.2.0",
+		"@tanstack/react-query": "^4.23.0",
+		"@trpc/client": "^10.9.0",
+		"@trpc/react-query": "^10.9.0",
+		"@trpc/server": "^10.9.0",
+		"prettier": "^2.8.3",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0"
+	},
+	"proxy": "http://localhost:8080/"
 }
