@@ -1,6 +1,7 @@
-const autoprefixer = require("autoprefixer");
-const tailwindcss = require("tailwindcss");
+/* eslint-disable */
+const autoprefixer = require('autoprefixer');
+const tailwindcss = require('tailwindcss');
 
 module.exports = {
-  plugins: [tailwindcss, autoprefixer],
+	plugins: [tailwindcss, autoprefixer],
 };
