@@ -1,64 +1,76 @@
-const HtmlWebPackPlugin = require("html-webpack-plugin");
-const ModuleFederationPlugin = require("webpack/lib/container/ModuleFederationPlugin");
+/* eslint-disable */
+const HtmlWebPackPlugin = require('html-webpack-plugin');
+const ModuleFederationPlugin = require('webpack/lib/container/ModuleFederationPlugin');
 
-const deps = require("./package.json").dependencies;
+const deps = require('./package.json').dependencies;
 module.exports = {
-  output: {
-    publicPath: "http://localhost:3000/",
-  },
+	output: {
+		publicPath: 'http://localhost:3000/',
+	},
 
-  resolve: {
-    extensions: [".tsx", ".ts", ".jsx", ".js", ".json"],
-  },
+	resolve: {
+		extensions: ['.tsx', '.ts', '.jsx', '.js', '.json'],
+	},
 
-  devServer: {
-    port: 3000,
-    historyApiFallback: true,
-  },
+	devServer: {
+		port: 3000,
+		historyApiFallback: true,
+	},
 
-  module: {
-    rules: [
-      {
-        test: /\.m?js/,
-        type: "javascript/auto",
-        resolve: {
-          fullySpecified: false,
-        },
-      },
-      {
-        test: /\.(css|s[ac]ss)$/i,
-        use: ["style-loader", "css-loader", "postcss-loader"],
-      },
-      {
-        test: /\.(ts|tsx|js|jsx)$/,
-        exclude: /node_modules/,
-        use: {
-          loader: "babel-loader",
-        },
-      },
-    ],
-  },
+	module: {
+		rules: [
+			{
+				test: /\.m?js/,
+				type: 'javascript/auto',
+				resolve: {
+					fullySpecified: false,
+				},
+			},
+			{
+				test: /\.(css|s[ac]ss)$/i,
+				use: ['style-loader', 'css-loader', 'postcss-loader'],
+			},
+			{
+				test: /\.(ts|tsx|js|jsx)$/,
+				exclude: /node_modules/,
+				use: {
+					loader: 'babel-loader',
+				},
+			},
+			{
+				test: /\.(jpe?g|gif|png|svg)$/i,
+				use: [
+					{
+						loader: 'url-loader',
+						options: {
+							limit: 10000,
+						},
+					},
+				],
+			},
+		],
+	},
 
-  plugins: [
-    new ModuleFederationPlugin({
-      name: "client",
-      filename: "remoteEntry.js",
-      remotes: {},
-      exposes: {},
-      shared: {
-        ...deps,
-        react: {
-          singleton: true,
-          requiredVersion: deps.react,
-        },
-        "react-dom": {
-          singleton: true,
-          requiredVersion: deps["react-dom"],
-        },
-      },
-    }),
-    new HtmlWebPackPlugin({
-      template: "./src/index.html",
-    }),
-  ],
+	plugins: [
+		new ModuleFederationPlugin({
+			name: 'client',
+			filename: 'remoteEntry.js',
+			remotes: {},
+			exposes: {},
+			shared: {
+				...deps,
+				react: {
+					singleton: true,
+					requiredVersion: deps.react,
+				},
+				'react-dom': {
+					singleton: true,
+					requiredVersion: deps['react-dom'],
+				},
+			},
+		}),
+		new HtmlWebPackPlugin({
+			template: './src/index.html',
+		}),
+	],
 };
