@@ -1,10 +1,10 @@
-import express from "express";
+import express from 'express';
 import * as trpcExpress from '@trpc/server/adapters/express';
-import { createContext } from "./context";
-import { PrismaClient } from '@prisma/client'
-import cors from "cors";
-import * as dotenv from "dotenv";
-import { appRouter } from "./routers/_app";
+import { createContext } from './context';
+import { PrismaClient } from '@prisma/client';
+import cors from 'cors';
+import * as dotenv from 'dotenv';
+import { appRouter } from './routers/_app';
 
 dotenv.config();
 
@@ -13,13 +13,15 @@ export const prisma = new PrismaClient();
 const app = express();
 app.use(cors());
 
-app.use('/trpc', trpcExpress.createExpressMiddleware({
-  router: appRouter,
-  createContext
-})
+app.use(
+	'/trpc',
+	trpcExpress.createExpressMiddleware({
+		router: appRouter,
+		createContext,
+	})
 );
 
 const port = 8080;
 app.listen(port, () => {
-  console.log(`server listening at http://localhost:${port}`);
+	console.log(`server listening at http://localhost:${port}`);
 });
