@@ -9,16 +9,14 @@ export const publicProcedure = t.procedure;
 export const mergeRouters = t.mergeRouters;
 
 export const isAuthed = middleware(async ({ next, ctx }) => {
-    if (!ctx.user) {
-      throw new TRPCError({ code: 'UNAUTHORIZED' });
-    }
-    return next({
-      ctx: {
-        user: ctx.user,
-      }
-    })
-  })
+	if (!ctx.user) {
+		throw new TRPCError({ code: 'UNAUTHORIZED' });
+	}
+	return next({
+		ctx: {
+			user: ctx.user,
+		},
+	});
+});
 
 export const protectedProcedure = publicProcedure.use(isAuthed);
-
-
